fix(libslirp): Suppress warnings and fix implicit casts in libslirp code
Silence warnings for third-party code on Apple platforms and correct implicit casts to prevent potential issues.

+++ src/core_melonds/src/net/libslirp/src/tftp.c
@@ -136,7 +136,7 @@ static int tftp_read_data(struct tftp_session *spt, uint32_t block_nr,
             return -1;
         }
 
-        bytes_read = read(spt->fd, buf, len);
+        bytes_read = (int)read(spt->fd, buf, len);
     }
 
     return bytes_read;
@@ -232,8 +232,8 @@ static void tftp_send_error(struct tftp_session *spt, uint16_t errorcode,
     slirp_pstrcpy((char *)tp->x.tp_error.tp_msg, sizeof(tp->x.tp_error.tp_msg),
                   msg);
 
-    m->m_len = sizeof(struct tftp_t) - (TFTP_BLOCKSIZE_MAX + 2) + 3 +
-               strlen(msg) - sizeof(struct udphdr);
+    m->m_len = (int)(sizeof(struct tftp_t) - (TFTP_BLOCKSIZE_MAX + 2) + 3 +
+               strlen(msg) - sizeof(struct udphdr));
     tftp_udp_output(spt, m, &recv_tp->hdr);
 
 out:
@@ -397,7 +397,7 @@ static void tftp_handle_rrq(Slirp *slirp, struct sockaddr_storage *srcsas,
 
             if (tsize == 0) {
                 if (stat(spt->filename, &stat_p) == 0)
-                    tsize = stat_p.st_size;
+                    tsize = (int)stat_p.st_size;
                 else {
                     tftp_send_error(spt, 1, "File not found", tp);
                     return;
